+++ CHANGES.md
@@ -84,6 +84,10 @@ CHANGES
     * Added `accessTokenResources`.
     * Added `resources`.
 
+- `TokenUpdateRequest` class
+    * Added `certificateThumbprint`.
+    * Added `dpopKeyThumbprint`.
+
 
 1.0.2 (2019-08-09)
 ------------------